+++ onsenui/esm/elements/ons-tabbar.spec.js
@@ -120,12 +120,17 @@ describe('OnsTabbarElement', () => {
     });
   });
 
-  it('sets tab initial index if none is provided', () => {
+  it('sets tab initial index if none is provided', done => {
     let element = ons.createElement(`<ons-tabbar><ons-tab label="t1"></ons-tab></ons-tabbar>`);
     expect(element.tabs[0].hasAttribute('active')).to.be.true;
 
-    element = ons.createElement(`<ons-tabbar activeIndex="1"><ons-tab label="t1"></ons-tab><ons-tab label="t2"></ons-tab></ons-tabbar>`);
-    expect(element.tabs[1].hasAttribute('active')).to.be.true;
+    element = ons.createElement(`<ons-tabbar active-index="1"><ons-tab label="t1"></ons-tab><ons-tab label="t2"></ons-tab></ons-tabbar>`);
+    document.body.appendChild(element);
+    setImmediate(() => {
+      expect(element.tabs[1].hasAttribute('active')).to.be.true;
+      element.remove();
+      done();
+    });
   });
 
   describe('#setTabbarVisibility()', () => {
@@ -344,6 +349,34 @@ describe('OnsTabbarElement', () => {
       element.onSwipe = assert.fail;
       element.remove();
       element.dispatchEvent(new Event('swipe'));
+    });
+  });
+
+  describe('#hideTabs', () => {
+    it('hides the tabbar when set', done => {
+      const element = ons._util.createElement('<ons-tabbar hide-tabs></ons-tabbar>');
+      document.body.appendChild(element);
+      setImmediate(() => {
+        const tabs = element.querySelector('.tabbar');
+        expect(tabs.style.display).to.equal('none');
+        element.remove();
+        done();
+      });
+    });
+  });
+
+  describe('#activeIndex', () => {
+    it("doesn't throw an error when the tabbar is not connected", () => {
+      const element = ons._util.createElement(`
+        <ons-page>
+          <ons-tabbar active-index="2">
+            <ons-tab page="page"></ons-tab>
+            <ons-tab page="page"></ons-tab>
+            <ons-tab page="page"></ons-tab>
+          </ons-tabbar>
+          <template id="page">page</template>
+        </ons-page>
+      `);
     });
   });
 });